Moved ChatPanel to separate file
