+++ react/gameday2/gameday2.js
@@ -5,6 +5,7 @@ var _ = require('underscore');
 
 import GamedayNavbar from './components/GamedayNavbar';
 import VideoGrid from './components/VideoGrid';
+import ChatPanel from './components/ChatPanel';
 
 var GamedayFrame = React.createClass({
   getInitialState: function() {
@@ -128,21 +129,6 @@ var GamedayFrame = React.createClass({
   },
   handleWebcastReset: function() {
     this.setState({displayedWebcasts: []});
-  }
-});
-
-var ChatPanel = React.createClass({
-  render: function() {
-    var classes = classNames({
-      'hidden': !this.props.enabled,
-      'pull-right': true,
-      'sidebar': true,
-    });
-    return (
-      <div className={classes}>
-        <iframe frameBorder="0" scrolling="no" id="chat_embed" src="http://twitch.tv/chat/embed?channel=tbagameday&amp;popout_chat=true" height="100%" width="100%"></iframe>
-      </div>
-    );
   }
 });
